[BE] refactor: 외부 저장소 (S3) 관련 로직 분리 #745
- Storage 컨트롤러와 서비스 추가
- 기존 메서드 이동
