+++ backend/admin/src/main/java/com/onair/hearit/admin/infrastructure/s3/FileStorage.java
@@ -5,18 +5,40 @@ import com.onair.hearit.admin.exception.custom.AdminInvalidInputException;
 import com.onair.hearit.core.domain.FileType;
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.URL;
+import java.time.Duration;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.multipart.MultipartFile;
 import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 import software.amazon.awssdk.services.s3.model.S3Exception;
+import software.amazon.awssdk.services.s3.presigner.S3Presigner;
+import software.amazon.awssdk.services.s3.presigner.model.PutObjectPresignRequest;
 
 @RequiredArgsConstructor
 public class FileStorage {
 
+    private static final Duration PUT_PRESIGNED_URL_DURATION = Duration.ofMinutes(3);
+
     private final S3Client s3Client;
+    private final S3Presigner s3Presigner;
     private final String bucket;
+
+    public URL createPutUrl(String key) {
+        key = validateKey(key);
+        PutObjectRequest put = PutObjectRequest.builder()
+                .bucket(bucket)
+                .key(key)
+                .build();
+
+        PutObjectPresignRequest presign = PutObjectPresignRequest.builder()
+                .putObjectRequest(put)
+                .signatureDuration(PUT_PRESIGNED_URL_DURATION)
+                .build();
+
+        return s3Presigner.presignPutObject(presign).url();
+    }
 
     public String uploadFile(MultipartFile multipartFile, FileType fileType) {
         String filePath = fileType.getUploadPath() + multipartFile.getOriginalFilename();
@@ -24,8 +46,8 @@ public class FileStorage {
     }
 
     private String uploadToS3(MultipartFile multipartFile, String filePath) {
-        String key = validateKey(filePath);
         try (InputStream inputStream = multipartFile.getInputStream()) {
+            String key = validateKey(filePath);
             s3Client.putObject(PutObjectRequest.builder()
                     .bucket(bucket)
                     .key(key)
