+++ backend/admin/src/main/java/com/onair/hearit/admin/presentation/api/AdminHearitController.java
@@ -2,20 +2,15 @@ package com.onair.hearit.admin.presentation.api;
 
 import com.onair.hearit.admin.application.AdminHearitService;
 import com.onair.hearit.admin.dto.request.AdminPagingRequest;
-import com.onair.hearit.admin.dto.request.HearitFileUpdateRequest;
 import com.onair.hearit.admin.dto.request.HearitInfoUpdateRequest;
 import com.onair.hearit.admin.dto.request.HearitMetaDataRequest;
-import com.onair.hearit.admin.dto.request.PresignedUrlRequest;
 import com.onair.hearit.admin.dto.response.AdminHearitResponse;
 import com.onair.hearit.admin.dto.response.AdminPagedResponse;
-import com.onair.hearit.admin.dto.response.FilesPresignedUrlResponse;
-import com.onair.hearit.core.domain.FileType;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -40,13 +35,6 @@ public class AdminHearitController {
         return ResponseEntity.ok(response);
     }
 
-    @PostMapping("/presigned-url")
-    public ResponseEntity<FilesPresignedUrlResponse> createPresignedUrl(
-            @RequestBody @Valid PresignedUrlRequest request) {
-        FilesPresignedUrlResponse response = adminHearitService.getFilesPresignedUrl(request);
-        return ResponseEntity.ok(response);
-    }
-
     @PostMapping
     public ResponseEntity<Void> createHearit(@RequestBody @Valid HearitMetaDataRequest request) {
         adminHearitService.addHearitMetaData(request);
@@ -58,30 +46,6 @@ public class AdminHearitController {
             @PathVariable Long hearitId,
             @RequestBody @Valid HearitInfoUpdateRequest request) {
         adminHearitService.modifyHearitMetaData(hearitId, request);
-        return ResponseEntity.noContent().build();
-    }
-
-    @PutMapping("/{hearitId}/original-audio")
-    public ResponseEntity<Void> updateHearitOriginalAudio(
-            @PathVariable Long hearitId,
-            @ModelAttribute @Valid HearitFileUpdateRequest request) {
-        adminHearitService.modifyHearitFile(hearitId, request, FileType.ORIGINAL);
-        return ResponseEntity.noContent().build();
-    }
-
-    @PutMapping("/{hearitId}/short-audio")
-    public ResponseEntity<Void> updateHearitShortAudio(
-            @PathVariable Long hearitId,
-            @ModelAttribute @Valid HearitFileUpdateRequest request) {
-        adminHearitService.modifyHearitFile(hearitId, request, FileType.SHORT);
-        return ResponseEntity.noContent().build();
-    }
-
-    @PutMapping("/{hearitId}/script")
-    public ResponseEntity<Void> updateHearitScript(
-            @PathVariable Long hearitId,
-            @ModelAttribute @Valid HearitFileUpdateRequest request) {
-        adminHearitService.modifyHearitFile(hearitId, request, FileType.SCRIPT);
         return ResponseEntity.noContent().build();
     }
 }
